analysis/neural_category_predictions/02_task_classification_model/pyproject.toml
[project]
name = "02-task-classification-model"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "model2vec[train] (>=0.6.0,<0.7.0)",
    "scikit-learn (>1.6.1)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pathlib (>=1.0.1,<2.0.0)",
    "torch (==2.1.0)",
    "numpy (>=1.24,<2)",
    "pandas (>=2.3.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
